Refactored server.py into several files to make easier understanding
