+++ apps/algorithm/search/src/config.py
@@ -1,0 +1,7 @@
+import os
+
+DEFAULT_INDEX = "listings"
+DISTANCE_TO_SEARCH_WITHIN = "5km"
+
+ES_ENDPOINT = os.getenv("ES_ENDPOINT")
+DB_ENDPOINT = os.getenv("DB_ENDPOINT")
